Add files via upload
Library file. This is where all of the possible words are stored, broken up by a new line for each word.
This is a default list of up to 10,000 words and abbreviations, but can be easily replaced with your own word list by savings over the contents of the file.
This file is required to run the python project.