Add files via upload
This file is not necessary, but you will need to have a similar file if you want to import your own boggle board. This file thus serves as an example to use to base your own on.
The board must be a X by X dimensions, or it will not properly work.
The board is broken up by rows of columns of the same dimensions where each is separated by a space.
Example:
a b e
c d f
g h i
This is a 3-by-3 board.
And the words you can find must be connected horizontally, vertically, or diagonally. 
Example word: 
bed
cad